Correct variable shadowing/capturing within block scopes
Variables captured by closures within block scopes will now be correctly
resolved even when they are later shadowed within the same scope.

Previously, the following code would print 4 and then 6. It now
correctly prints 4 twice.
```
{
  var a = 4;
  fun foo() {
    print a;
  }

  foo();
  var a = 6;
  foo();
}
```

Note that in the global scope, shadowing a variable is equivalent
to reassigning it; this is to support REPL usage, where it is
often desirable to redeclare functions and variables and have the
redeclaration take effect in previous definitions. The following code
correctly prints 4 and then 6:

```
var a = 4;
fun foo() {
  print a;
}

foo();
var a = 6;
foo();
```